Another windows compilation warning fixed
git-svn-id: http://svn.cern.ch/guest/lhcb/Rec/trunk@89832 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 2967858577ecdac3cc31d1b7c7a1abc72bb1f89f

+++ Rec/Tf/FastVelo/src/FastVeloHit.h
@@ -105,7 +105,7 @@ public:
   void   setGlobal( double r )  { m_global = float(r); }
   void   setUsed()              { m_nbUsed++; }
   void   clearUsed()            { m_nbUsed--; }
-  void   setGlobalPosition( float x, float y ) { 
+  void   setGlobalPosition( double x, double y ) { 
     m_xGlobal = float(x); 
     m_yGlobal = float(y);
   }
